% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding.R
\name{b_binding}
\alias{b_binding}
\title{Calculate Binding Affinities}
\usage{
b_binding(
  data,
  conc,
  resp,
  ...,
  model = drc::LL.4(names = c("slope", "min", "max", "kd"))
)
}
\arguments{
\item{data}{data.frame or tibble containing dose response data.}

\item{conc}{Column with the concentration information.}

\item{resp}{Column containing the response information.}

\item{...}{Columns that contain the grouping information for different
samples.}

\item{model}{Model to fit. Defaults to \code{drc::LL.4()} for log-logistic
dose-response model for binding.}
}
\value{
A \code{tibble()} with list-columns containing the data, the predictions,
the residuals and the coefficients of the model.
}
\description{
Calculated binding affinities (Kd) from experimental data that contains dose
and response information. Could be titrated concentration and resulting
fluorescence, or percentage shift in a EMSA or something similar.
}
