% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{b_params}
\alias{b_params}
\title{Extract Model Parameters to Columns}
\usage{
b_params(data)
}
\arguments{
\item{data}{Tibble containing a `coefs` column resulting from a model fit.}
}
\value{
tibble containing grouped variables and extracted parameters.
}
\description{
Extract Model Parameters to Columns
}
\examples{


}
