% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{b_plot}
\alias{b_plot}
\title{Plot Data with Fitted  Models}
\usage{
b_plot(data, ..., facet = TRUE)
}
\arguments{
\item{data}{Tibble with nested columns of \code{data}, \code{drmod}, \code{line}, \code{coefs}
made through one of the \code{b_binding()} or \verb{b_enzyme_*()} functions.
currently used facetting variables.}

\item{...}{Additional \code{ggplot2::aes()} parameters to be applied to the plot.}

\item{facet}{Logical, whether to apply \code{ggplot2::facet_wrap()} based on the}
}
\value{
\code{ggplot2::ggplot()} object.
}
\description{
Plot Data with Fitted  Models
}
\examples{
library(tidyverse)
Puromycin \%>\%
  b_enzyme_rate(conc, rate, state) \%>\%
  b_plot()
}
