% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding.R
\name{bio_binding}
\alias{bio_binding}
\title{Calculate Binding Affinities}
\usage{
bio_binding(data, conc, resp, group = NULL, slope = 1)
}
\arguments{
\item{data}{data.frame or tibble containing dose response data.}

\item{conc}{Column with the concentration information.}

\item{resp}{Column containing the response information.}

\item{group}{allowed to be free (set \code{slope = NA}).}

\item{slope}{For all 1:1 binding interactions, the slope for the log -
logistic curve should be 1. If the interaction has some cooperativity or it
is not a 1:1 binding interaction, then this value will change and should be}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Calculated binding affinities (Kd) from experimental data that contains dose
and response information. Could be titrated concentration and resulting
fluorescence, or percentage shift in a EMSA or something similar.
}
\examples{
Puromycin \%>\%
  bio_binding(conc, rate, group = state)
}
