% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{bio_coefs}
\alias{bio_coefs}
\title{Extract Model Parameters to Columns}
\usage{
bio_coefs(data)
}
\arguments{
\item{data}{Tibble containing a \code{coefs} column resulting from a model fit.}
}
\value{
tibble containing grouped variables and extracted parameters.
}
\description{
Extract Model Parameters to Columns
}
\examples{

# Fitting MM models to the enzymatic data, pull out the relevant information.
Puromycin \%>\%
  bio_enzyme_rate(conc, rate, group = state) \%>\%
  bio_coefs()
}
