% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enzyme_rate.R
\name{bio_enzyme_rate}
\alias{bio_enzyme_rate}
\title{Fit a Michaelis-Menton Curve to Enzymatic Data}
\usage{
bio_enzyme_rate(data, conc, rate, group = NULL, min = 0)
}
\arguments{
\item{data}{Data frame with columns for concentration, reaction rate and
optionally grouping information.}

\item{conc}{Column containing the concentration data.}

\item{rate}{Column containing the enzyme rate data.}

\item{group}{}

\item{min}{Minimum value of enzyme rate. Defaults to 0, as if there is 0
substrate there should be no enzymatic activity in a properly blanked
experiment. Set to \code{NA} to allow the model to fit the minimum value.}
}
\value{
\code{tibble} with nested list columns of data, model, predictions,
residuals and coefficients.
}
\description{
This function fits a Michaelis-Menton model to the dose-response data of
enzymatic reactions. By default the minimum rate is locked to 0, if you wish
to let the function find the minimum (if enzyme rate is > 0 at concentration
0) then supply the \code{min = NA} or manually set it i.e. \code{min = 2.5}.
}
\examples{
# Fitting MM curves to the enzymatic data inside of datasets::Puromycin
library(bicohemr)
df <- Puromycin \%>\%
  bio_enzyme_rate(conc, rate, state)

# the result is a tibble with a column for the data, a column for the
# calculated # model and a column for the relevant coefficients extracted from
# the model
df

# you can extract the coefficient data using either `bio_coefs()` or
# `unnest()` on the column
bio_coefs(df)
tidyr::unnest(df, coefs)

# quick plot can be made for inspecing the results with `bio_plot()`
df \%>\%
  bio_plot()
}
