% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bio_plot}
\alias{bio_plot}
\title{Plot Data with Fitted  Models}
\usage{
bio_plot(
  data,
  ...,
  facet = TRUE,
  line_col = "black",
  line_type = "solid",
  line_size = 0.8,
  line_alpha = 1,
  point_shape = 21,
  point_col = "black",
  point_alpha = 1,
  point_size = 1
)
}
\arguments{
\item{data}{Tibble with nested columns of \code{data}, \code{drmod}, \code{line}, \code{coefs}
made through one of the \code{bio_binding()} or \verb{bio_enzyme_*()} functions.}

\item{...}{Additional \code{ggplot2::aes()} parameters to be applied to the plot.}

\item{facet}{Logical, whether to apply \code{ggplot2::facet_wrap()} based on the}

\item{point_size}{currently used facetting variables.}
}
\value{
\code{ggplot2::ggplot()} object.
}
\description{
Plot Data with Fitted  Models
}
\examples{
library(biochemr)
Puromycin \%>\%
  bio_enzyme_rate(conc, rate, group = state) \%>\%
  bio_plot()
}
