% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bio_plot_coefs}
\alias{bio_plot_coefs}
\title{Comparison Plot of Estimated Terms}
\usage{
bio_plot_coefs(data, terms = NULL, ..., pointSize = 3, errorBarHeight = 0.3)
}
\arguments{
\item{data}{Nested \code{\link{tibble}} that is the result from one of the fitting functions.}

\item{...}{Additional arguments to be passed to \code{ggplot2::aes()} such as
\code{colour = sample}.}

\item{pointSize}{Size of points, defaults to \code{3}.}

\item{errorBarHeight}{Relative height of the errorbars, defaults to \code{0.3}.}

\item{term}{Term to be plotted as string, i.e. \code{"kd"} or \code{"rate"}.}

\item{group}{Column that results are grouped by.}
}
\value{
\code{\link{ggplot}} object
}
\description{
Comparison Plot of Estimated Terms
}
\details{
Creates a dot-and-whisker plot for quick comparison of estimated
terms, also showing the calculated standard errors as errorbars.
}
\examples{
library(biochemr)

DNase \%>\%
  dplyr::group_by(Run) \%>\%
  bio_binding(conc, density) \%>\%
  bio_plot_coefs(Run, term = "kd", colour = Run)
}
