% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response.R
\name{b_dose_resp}
\alias{b_dose_resp}
\title{Fitting of a Dose Response Model}
\usage{
b_dose_resp(
  data,
  dose,
  response,
  ...,
  model = drc::LL.4(names = c("slope", "min", "max", "ec50"))
)
}
\arguments{
\item{dose}{Column name for the dose data.}

\item{response}{Column name for the response data.}

\item{...}{Columns by which to group the data before fitting models.}

\item{model}{Model function from \code{{drc}} to be fit to the data. Defaults to
\code{drc::LL.4()} for the \code{b_dose_resp()} function.}

\item{.data}{Data frame contain columns for dose, response and if required
grouping variables.}
}
\value{
A \code{tibble()} with list-columns containing the data, the predictions,
the residuals and the coefficients of the model.
}
\description{
Fitting of a Dose Response Model
}
