% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enzyme_rate.R
\name{b_enzyme_rate}
\alias{b_enzyme_rate}
\title{Fit Michaelis Menton Curve}
\usage{
b_enzyme_rate(
  data,
  conc,
  rate,
  ...,
  model = drc::MM.2(names = c("Vmax", "Km"))
)
}
\arguments{
\item{data}{Data frame with columns for concentration, reaction rate and
optionally grouping information.}

\item{conc}{column with concentration data.}

\item{rate}{Column with reaction rate data.}

\item{...}{Columns with grouping data.}

\item{model}{Model to fit defaults to `drc::MM.2()`}
}
\value{
`tibble` with nested list columns of data, model, predictions,
 residuals and coefficients.
}
\description{
Fit Michaelis Menton Curve
}
