% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{b_plot}
\alias{b_plot}
\title{Plot Data with Fitted  Models}
\usage{
b_plot(.data, ..., facet = TRUE)
}
\arguments{
\item{.data}{Data frame with nested columns of data, models and predictions.}

\item{...}{Additional \code{ggplot2::aes()} parameters to be applied to the plot.}

\item{facet}{Logical, whether to apply \code{ggplot2::facet_wrap()} based on the
currently used facetting variables.}
}
\value{
\code{ggplot2::ggplot()} object.
}
\description{
Plot Data with Fitted  Models
}
\examples{

}
