% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fret.R
\name{fret}
\alias{fret}
\title{Calculate fret.}
\usage{
fret(donor, transfer)
}
\arguments{
\item{donor}{Donor fluorescent intensity.}

\item{transfer}{Transfer fluorescent intensity.}
}
\value{
Vector of calculated fret intensities.
}
\description{
A
}
